requirements.txt
opencv-python
scikit-image
pandas
pytesseract
jiwer
pillow